php7.1 warning fix
cfr. https://github.com/tubalmartin/YUI-CSS-compressor-PHP-port/pull/33

+++ cssmin.php
@@ -266,7 +266,7 @@ class CSSmin
 
 	// preserve flex, keeping percentage even if 0
 	$css = preg_replace_callback('/flex\s?:\s?((?:[0-9 ]*)\s?(?:px|em|auto|%)?(?:calc\(.*\))?)/i',array($this, 'replace_flex'),$css);
-		
+	
 	// Fix IE7 issue on matrix filters which browser accept whitespaces between Matrix parameters
 	$css = preg_replace_callback('/\s*filter\:\s*progid:DXImageTransform\.Microsoft\.Matrix\(([^\)]+)\)/', array($this, 'preserve_old_IE_specific_matrix_definition'), $css);
 
@@ -413,6 +413,7 @@ class CSSmin
         // restore preserved comments and strings in reverse order
         for ($i = count($this->preserved_tokens) - 1; $i >= 0; $i--) {
             $css = preg_replace('/' . self::TOKEN . $i . '___/', $this->preserved_tokens[$i], $css, 1);
+            // $css.=$this->preserved_tokens[$i];
         }
 
         // Trim the final string (for any leading or trailing white spaces)
@@ -470,8 +471,8 @@ class CSSmin
             if ($found_terminator) {
                 $token = $this->str_slice($css, $start_index, $end_index);
                 // remove whitespace, except if $token contains svg, which needs whitepace left as is
-                if (strpos($token,"<svg")===false && strpos($token,'svg+xml')===false) {
-	                $token = preg_replace('/\s+/', '', $token);
+				if (strpos($token,"<svg")===false && strpos($token,'svg+xml')===false) {
+					$token = preg_replace('/\s+/', '', $token);
                 }
                 $this->preserved_tokens[] = $token;
 
@@ -779,9 +780,9 @@ class CSSmin
     {
         if (is_string($size)) {
             switch (substr($size, -1)) {
-                case 'M': case 'm': return $size * 1048576;
-                case 'K': case 'k': return $size * 1024;
-                case 'G': case 'g': return $size * 1073741824;
+                case 'M': case 'm': (int) return $size * 1048576;
+                case 'K': case 'k': (int) return $size * 1024;
+                case 'G': case 'g': (int) return $size * 1073741824;
             }
         }
